savetradebot/requirements.txt
tensorflow-cpu==2.14.0
ta
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
sqlalchemy==2.0.36
matplotlib>=3.5.1
python-dotenv==1.0.0
ccxt==4.0.50


